A bit of refactoring
After some discussion with the assignment creator, I've moved a couple
of methods around for reading in the files, so that the 'stream' is
actually reading in another sentence.

Will continue to work on this, and get both 'streams' to be read at the
same time and work on getting them 'synced'.