Reading in both streams
Some refactoring.
Reading in from both 'streams' and syncing them up. Works halfway so
far, where Stream 1 is 'behind' Stream 2. I will be making the date/time
a little easier to compare, so we can find which Stream is behind
another Stream.

It feels wrong.. Reading Streams this way.. As with a real Stream, you
wouldn't hold off on reading one stream to search/wait for the other to
give you a timestamp matching the original stream, would you?

I guess you could, but then you're 'behind' on one stream, that while
you've been searching/waiting for the other stream to catch up, the
original stream is still sending data to you and incrementing it's
timestamp.

I feel this assignment is a little vague in what it's doing when
requesting syncing, or how to treat two static .dat text files as a
'stream', and should have been either explaining in more detail what
should or should not be allowed, or at least properly explain what is
meant by a 'stream', in this example, as there's not much way that two
text files, that are unchanging, can be properly treated as a stream,
'synced up' as if they were real.. etc.. Bah. Ambiguity is something
clients will give in the future, sure, but this is ambiguity without
real world evidence of what the intention is.

~end complaints about things I'll just deal with~