Inheritance FTW
Working on shared methods for inheritance with the sentences.

The last commit message got cut off and sent without allowing me to edit
it (using GitHubs own application.. That should be fixed.. Allow edits
of commit messages!)
Anyway, I was just stating how important it is that the ordering of the
lines of code are in correctly getting and declaring each stage of
reading in sentences and storing their info.

Speaking of.. The next step is getting the info into a space where it
can be called upon for the XML at each new timestamp :)