More work on data input
A bit more work on taking in data and saving it. Specifically the
Lat/Lng and Date/Time from GPGGAs, along with some checking for null
stuff. Fun times! I wonder how well this will translate to C and C++..

Will continue with the same thing tomorrow, taking in relevant data. I
need to start on checking the data integrity soon..
There's still a week and a half or so left to do this.. But oh so, so
much left to do..