More variables! Saving SNR values
Just more updates to the data as streams come in.
Now holding the SNRs of each satellite that comes in on a sentence that
actually has an SNR value.

A couple more sentences to go and then I can start the proper integrity
checking and XML output. The ambiguity of this NMEA standard is awful.
When we get new timestamps with lat/lng locations but no data to back it
up (i.e. GPGSVs with SNRs for the satellites), how can we be sure it's a
good signal? Do we go off the previouslly recorded satellite data, or
wait for the next data to say they are correct, which might not arrive
for another 1-3 seconds? It's not great.
I will go with the method of 'if it was okay a second okay, consider
this one okay', I believe. Seems the most logical, as we could be
waiting for another signal that isn't so great and could imply the
previous data was also invalid.
We'll see what happens when I reach checking for integrity of the data
and signal.