All sentences dealt with, some error fixes
The application now deals with each bit of sentence data that we're
interested in. The code needs some refactoring but it works as intended.

Did some error fixing with recording the Date. This NMEA format is so
bad.. Sending GPS with a time but no date on some sentences, but with
others having both. Blech. Anyway, got something working for it that
should suffice for this scope, but it's not perfect. If this were a
safety critical system, I'd probably recommend writing a new GPS
standard that wasn't so full of potholes...