Once Upon A Java!
The Java side is finished.
Output is complete. Horay! However, the route doesn't look perfect. I've
looked at Stream 1 alone, and it's smooth. I've looked at Stream 2
alone, and it's not smooth. It seems where route 1 is subbed for Stream
2, it goes a little wonky. Not too much I can do about that, I just hope
that is known by the assessors.

Either way, the program is functional and does as it should. Could
probably be refactored, but that may happen once I'm done with C or C++
and the documentation when I have the time available.