Do what you want because a Pirate is C
Reads in both 'streams' now, just like in Java.
Updates the time correctly, and can compare the times appropriately.
Looks for missing seconds, or if a stream is ahead/behind the other in
order to keep them in sync.

Updates to the Lat/Lng offsets and keeping track.

Next stage is:
Get the GPGSVs and the satellite SNRs.
Check for the 'best' stream using the SNRs.
Encode everything in XML.
Output to file.
Comment and refactor.