C++ Started and finished
C++ Program was started and completed.

Needs commenting, but it's otherwise as functional as the others.
I'm going to revisit the Java version and clean it up at a later date.
It's a little shoddy, could do with a massive clean and also isn't
precise enough on the Lat/Lng in some places. Looks like Fred took some
large steps in places he hasn't.