Lat/Lng NMEA to Decimal conversion updates
Updated the way that Lat/Lng is converted from NMEA to Decimal as I was
doing it incorrectly previously. It now works like a charm!

C++ is now commented and formatted.
C and C++ give some system console prints to demonstrate working
behaviour.
The screencasts for C and C++ and now done!