Heavily refactored Java
Changed a bunch in the Java structure and files. It now resembles the C
and C++ programs, and is a much better way of writing the application.

This is partially due to having a decent and updated assignment
specification now. The original was ambiguous and vague about what it
wanted, so I had more than I should have done in there (there's still
more than needed, but I'm happy with the result).