Preparing for assignment hand-in
Moved some files around for the assignment hand-in.
Created and written a README text file, documenting what files are being
handed in, where to find them and any specifics for building and running
the code,.

The content of this README has also been included in the README for
GitHub.

Final task before submission:
Write a (max 2 pages) document, detailing any assumptions or
simplifications I made, how I started and worked through the assignment
and conclude the end result.