Cleaned up the draft a bit: Draft 1.2 ish
Cleaned the text and structure a bit and added a new image for comparing
the 'Set Latitude' in C and in Java, for the clarity of the code they
are both written with.

I'll now add the cited references and then see if it's ready to submit
or needs more work.