Clean up. Final?
Did a clean up, and I'm at a stage where I'm okay to submit the current
version. I may come back to it before the deadline and go over it again
(I should do).

Next up is HCI and then continuing with the game and game report, so
there's still a lot to do. Plus web work for School spider. So we shall
see about my time constraints!